backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3
SQLAlchemy==2.0.23
Werkzeug==3.0.1
requests==2.31.0